// sp/src/app/pages/models/employee.model.ts
export class Employee {
    _id?: number = 0;
    firstname: string = '';
    lastname: string = '';
    birthdate: string = '';
    gender: string = '';
    education: string = '';
    company: string = '';
    jobExperience: number = 0;
    salary: number = 0;
    profile: string = '';
}